+++ ui/elements/options/slideroption.gd
@@ -14,6 +14,9 @@ func _ready() -> void:
 func set_label(text : String) -> void:
 	$BaseOption.set_label(text)
 
+func get_option_name() -> String:
+	return option_name
+
 func set_option_name(name : String) -> void:
 	option_name = name
 
@@ -21,8 +24,13 @@ func get_value() -> float:
 	return $BaseOption/HBoxContainer/HBoxContainer/HSlider.value
 
 func set_value(value : float) -> void:
+	$BaseOption/HBoxContainer/HBoxContainer/HSlider.disconnect("value_changed", self, "_on_HSlider_value_changed")
+	
 	$BaseOption/HBoxContainer/HBoxContainer/HSlider.value = value
-
+	$BaseOption/HBoxContainer/HBoxContainer/ValueDisplay.text = str(value)
+	
+	$BaseOption/HBoxContainer/HBoxContainer/HSlider.connect("value_changed", self, "_on_HSlider_value_changed")
+	
 func set_widget_node(path : Node) -> void:
 	widget_node = path