+++ widgets/hboxcontainer_options.gd
@@ -1,4 +1,7 @@
 extends Control
 
-func add_config_options(widget_id, config):
+func add_config_options(widget_id : String, config : Dictionary) -> void:
+	pass # This widget does not have any options
+
+func set_receiver_node(receiver : Node) -> void:
 	pass # This widget does not have any options